NEL: Add SQLitePersistentReportingAndNELStore
This CL adds SQLitePersistentReportingAndNELStore which persists
NEL policies to disk in a SQLite database. It will eventually also store
Reporting clients and reports.

The code is based on that of the SQLitePersistentCookieStore.

It is not hooked up to anything yet (just unittests).

Bug: 895821
Change-Id: I99ea14ace6f9b7fecd6075b1be98647db80d2bcb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1504493
Commit-Queue: Lily Chen <chlily@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#641361}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8c66199b51cc1f2c52f1b206823bc9054d6fa1bb