Pass in full Options to SetCanonicalCookieAsync
This will be used to check whether same site cookies are being set
in the proper context. This change alters all the callsites to
include whether it should be trusted to set such cookies or not,
so it's not a simple/mechanical transformation.

(The old bool for httponly gets passed inside the options object as well,
 like it would be for the get).

Bug: 837412

Change-Id: Ie7faf4e7cbc80706720e6ad20f6a46c759621f0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1471244
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Mohammad Refaat <mrefaat@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Jeremy Klein <jlklein@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#641695}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fdbc8beebe3e89c3e7c1a9ca2d6b039ba784fca4