Make ProxyResolvingClientSocket use ConnectJobs directly.
Previously, ProxyResolvingClientSocket created a socket by passing a
ClientSocketHandle over to a ClientSocketPool, which then created
a ConnectJob to handle establishing a connection.

This CL makes it so that ProxyResolvingClientSocket directly manages the
ConnectJob itself. The code doesn't get anything out of using a
ClientSocketPool, since the sockets aren't reuseable, and with the
socket pool refactor, it's possible to just use ConnectJobs directly.

This CL also works us towards a world in which a GroupId, along with
the choice of which ClientSocketPool to use, is sufficient to create
a ConnectJob, which will allow us to remove
ClientSocketPool::SocketParams, a pattern which has long been a source
of bugs.

Bug: 944911
Change-Id: Ia5e9b82dccd6f6ec8daa4f1900320ca9db4aeac7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1534647
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#644131}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e6d792f109806f540a985619866c41ad9e095da2