SimpleCache: fix entry leak in obscure backend-destruction circumstances
Those circumstances being the backend being deleted in between when the
entry creation/opening op has succeeded and when the callback was going
to be delivered to the client.

Also adjust the fuzzer to only commit the entry pointer to its table
after callback is invoked and not directly since it could end up trying
to close an open it never got a callback for (oh, and there was a likely
race with blockfile, too, since it writes out the pointer on a different
thread)

More evidence that this API should change, I think.

Bug: 946349
Change-Id: Ie17f23331e69eeae527918f181de787c8a550623
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1541667
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#646972}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 661d99dba5a2b4c080a9666259e0c020c81b7fd3