GetAdditionalErrorState refactor 4, add methods to get SSL information.
This CL adds two methods to ConnectJob to query whether an SSL error
occured, and to get the SSLCertRequestInfo, if there is one.
Previously this information could only be obtained by creating a
ClientSocketHandle and calling ConnectJob::SetAdditionalErrorState on
it.

This is part of an effort to remove ConnectJob::SetAdditionalErrorState,
which is a layering violation, since it takes a ClientSocketHandle,
which is a SocketPool object, and SocketPools are a higher layer than
ConnectJobs.

Bug: 927101
Change-Id: I9bca7cae89a0ff5a13c0fc217905b9861c0bcc4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1558586
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#649999}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6f84d1f1909fb497600a431026a6171937f2e024