Replace NetworkChangeNotifierPosix::NotifierThread with SequencedTaskRunner
Dedicated thread wastes resources and causes races in tests where
it can outlive thread pools used by DnsConfigServicePosix's
SerialWorkers.

This is a reland with fixed creation of SerialWorkers so they're created on
the thread they're deleted on, and a regression test that catches the original
reason for revert.  I'm also extending
DnsConfigServicePosixTest.DestroyWhileJobsWorking a bit to test more.
NotifyWatcherMac had to be converted to use the more modern
FileDescriptorWatcher so that it could run on sequences for the new test,
and this will also be useful when NetworkChangeNotifierMac has its Thread
converted to a SequencedTaskRunner.

Bug: 938126
Change-Id: I1a607001a8debf6730bdf8215a3985392291726b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1557698
Reviewed-by: Eric Orth <ericorth@chromium.org>
Commit-Queue: Paul Jensen <pauljensen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#650334}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9cb44c47cae5f5df81f78dd4d6e065418cfc0c83