Network Error Logging: Slight optimization to sampling if rate is 0 or 1
If the sampling rate is 0.0 or 1.0, there is no need to generate a
random number to see whether we should sample or not.
In most cases, this eliminates a file read to /dev/urandom which may
be slow.

Bug: None
Change-Id: I6ed8302574d45ddd1437610c8c5b03de187f8c85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1566112
Reviewed-by: Misha Efimov <mef@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#650449}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cac01ac7d49f8de4b119e5427b47adc3a477b68e