Check channel_ in WebSocket::OnBeforeSendHeadersComplete & OnHeadersReceivedComplete
The WebSocketChannel and URLRequest will be destroyed if the connection
is aborted, so the |out_headers| and |callback| in these methods would
not be safe to access.

Also use weakptr on
URLRequestHttpJob::NotifyBeforeStartTransactionCallback. This avoids an
issue where the callback is run while the URLRequest is cancelled but
not yet destroyed, similar to
https://crrev.com/be9d8e81e8c3ac82d25597e653d929dfff2d8ec1

Bug: 947407
Change-Id: I7d4e16859fa9f9c17ce4264eb78a9c4224a686b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1559493
Commit-Queue: Matt Mueller <mattm@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#650450}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 22372d439cafc4758fca7120fc5fe5df3825e9be