Rework HostResolutionCallback.
This is used to check for H2 sessions with the same IP and a matching
cert. This CL reworks it so that it only depends on information the
TransportConnectJob has available to it, so that it can be handled
entirely within the socket pools, without passing in a SocketParam,
which doesn't work correctly for a number of reasons (For instance,
if we have two requests, one that allows IP aliasing, and one that
does not, it's possible the IP aliasing request will cancel both
ConnectJobs with ERR_SPDY_SESSION_ALREADY_EXISTS).

Bug: 912727
Change-Id: I3526da33b027f953435373ec9d911dafb1f9e47b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1584414
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Bence Béky <bnc@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#655663}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5062be254c5abe25d86419f0ab11cb69ef3796f3