Remove unnecessary data copies in decoding data URLs.
In https://chromium-review.googlesource.com/c/chromium/src/+/1592099,
I reworked the UnescapeBinaryURLComponent API so that it couldn't use
the same string as input and output. In some cases, that added a copy
from one buffer to another, instead of a copy within a buffer. The only
potentially concerning case that affects is the case of data URLs, as
they can be quite large.

This CL remove an entire copy (Not just replacing a copy-between-buffers
with a copy-within-a-buffer) in the case of Base64 and text-encoded data
URLs, by introducing the use of StringPiece to wrap just the body part
of a data URL. For other types of data URLs, this CL does not save any
copies.

Bug: 849998
Change-Id: I9c141c80a26ba620745d2cefded84819c9b59cc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1594732
Reviewed-by: David Benjamin <davidben@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#656554}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 57fd88b66be7546725a8ef689ed3b9afa3f9a853