[net/auth] HttpAuthController assumes URL is valid.
Setting the auth_path_ field of HttpAuthController was intended to be
set to an empty string if the target of authentication was a proxy. Due
to a typo, this branch was never taken. Instead HttpAuthController uses
the full URL as provided by the caller.

This change removes the unused condition and assumes that the input URL
is already well formed.

Bug: 927182
Change-Id: I79211e9c9e071dab8730255c56ca95409f8d7f12
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1597109
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Asanka Herath <asanka@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#657584}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d856bd016113f3ba921c5c36dd9a17d93045f345