Make AuthCallback explicitly RepeatingCallback, use BindRepeating.
The use of Bind() is deprecated, replace with BindRepeating because
this callback is called more than once, and make the type explicit
as well in the receiver.

This was discussed in "Should AuthCallback be a OnceCallback?" on
blink-network-dev@chromium.org.

R=mmenke@chromium.org

Bug: 953861
Change-Id: Iaa4a155e767da077688a4a107d2f0d7bf37d6a15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1615835
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#661007}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c988f0be66bc36c20d14ce2ebeabbf511f208d95