Unwind the CT DNS-based proof inclusion experiment
Chromium has had experimental support for the CT-over-DNS draft at
https://github.com/google/certificate-transparency-rfcs/blob/master/dns/draft-ct-over-dns.md
for some time. In support of this experiment, it's received STHes from
Google using the Component Updater, and when a fresh STH is received,
used that to trigger checks for any SCTs it may have from that Log, to
ensure the Log is well-behaving.

Due to the use of TXT records, this code was only implemented using our
built-in resolver code, and only on platforms where CT was enabled;
effectively limiting it to !Windows and !Mobile.

Many of the privacy assumptions around this code were conceived of
prior to, and independent of, work such as DoH or ESNI, and may not
hold the same. Further, the limited experiments performed suggest
abnormally high failure rates, which will require significant time
to dig into.

As such, we're putting a hold on the experiment to focus on other
higher priority tasks. Due to the complexity this introduces for
reasoning about DoH and ESNI, and with a mind to binary size, rather
than leave the code in place, this CL unwinds it all, as hermetically
as possible, so that it can be revisited in the future.

Bug: 506227
Change-Id: I1b477b34e4445bc06cc4cef95b92909e74d28327
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1582766
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Misha Efimov <mef@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Ryan Sleevi <rsleevi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#664012}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3e59ea3a0318aa866a574edac1d6c770e1dee504